added code for functional integration
